backend/requirements.txt
pymongo==3.10
Flask==1.1.1
Flask-Cors==3.0.8
python-telegram-bot
